fix(adapter): auto 워크플로우 생성 정합성 복구
OpenCode가 얇은 prompt entrypoint를 skill로 노출하면서 실제 orchestration 의미를 잃고 있었고, Codex/Gemini도 router·canary surface drift가 생길 수 있어 플랫폼별 생성 경로를 다시 정렬합니다.
이 변경은 하네스 바이너리 업데이트만으로 각 플랫폼의 auto workflow가 기대한 native surface를 복구하도록 만듭니다.

Constraint: 플랫폼별 native workflow surface를 유지하고 prompt와 skill의 역할을 섞지 않는다
Rejected: Claude router 재사용에 추가 문자열 치환만 덧대는 접근
Confidence: medium
Scope-risk: module
Reversibility: moderate
Directive: root harness generated output sync와 worker flake 분리는 후속 변경으로 유지한다
Tested: go test ./pkg/adapter/codex ./pkg/adapter/gemini ./pkg/adapter/opencode ./pkg/content; auto check --arch --quiet --staged
Not-tested: go test ./... 는 darwin 환경에서 pkg/worker TestExecuteWithParallel_DoesNotStartExpiredQueuedTask 실패로 녹색이 아님
Related: v0.40.13 release

🐙 Autopus <noreply@autopus.co>
