+++ pkg/adapter/codex/codex_standard_skills.go
@@ -1,16 +1,12 @@
 package codex
 
 import (
-	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/insajin/autopus-adk/pkg/adapter"
 	"github.com/insajin/autopus-adk/pkg/config"
-	pkgcontent "github.com/insajin/autopus-adk/pkg/content"
-	"github.com/insajin/autopus-adk/templates"
 )
 
 type standardSkillSpec struct {
@@ -55,67 +51,6 @@ var codexUserFacingSkills = []standardSkillSpec{
 		Description:  "Autopus 배포 검증 워크플로우. build, E2E, 브라우저 건강 검진을 실행할 때 사용합니다.",
 		TemplatePath: "codex/skills/auto-canary.md.tmpl",
 	},
-}
-
-type pluginManifest struct {
-	Name        string          `json:"name"`
-	Version     string          `json:"version"`
-	Description string          `json:"description"`
-	Author      pluginAuthor    `json:"author"`
-	Homepage    string          `json:"homepage"`
-	Repository  string          `json:"repository"`
-	License     string          `json:"license"`
-	Keywords    []string        `json:"keywords"`
-	Skills      string          `json:"skills"`
-	Interface   pluginInterface `json:"interface"`
-}
-
-type pluginAuthor struct {
-	Name  string `json:"name"`
-	Email string `json:"email"`
-	URL   string `json:"url"`
-}
-
-type pluginInterface struct {
-	DisplayName       string   `json:"displayName"`
-	ShortDescription  string   `json:"shortDescription"`
-	LongDescription   string   `json:"longDescription"`
-	DeveloperName     string   `json:"developerName"`
-	Category          string   `json:"category"`
-	Capabilities      []string `json:"capabilities"`
-	WebsiteURL        string   `json:"websiteURL"`
-	PrivacyPolicyURL  string   `json:"privacyPolicyURL"`
-	TermsOfServiceURL string   `json:"termsOfServiceURL"`
-	DefaultPrompt     []string `json:"defaultPrompt"`
-	BrandColor        string   `json:"brandColor"`
-}
-
-type marketplaceDoc struct {
-	Name      string             `json:"name"`
-	Interface marketplaceUI      `json:"interface,omitempty"`
-	Plugins   []marketplaceEntry `json:"plugins"`
-}
-
-type marketplaceUI struct {
-	DisplayName string `json:"displayName,omitempty"`
-}
-
-type marketplaceEntry struct {
-	Name     string            `json:"name"`
-	Source   marketplaceSource `json:"source"`
-	Policy   marketplacePolicy `json:"policy"`
-	Category string            `json:"category"`
-}
-
-type marketplaceSource struct {
-	Source string `json:"source"`
-	Path   string `json:"path"`
-}
-
-type marketplacePolicy struct {
-	Installation   string   `json:"installation"`
-	Authentication string   `json:"authentication"`
-	Products       []string `json:"products,omitempty"`
 }
 
 func (a *Adapter) renderStandardSkills(cfg *config.HarnessConfig) ([]adapter.FileMapping, error) {
@@ -218,347 +153,6 @@ func (a *Adapter) preparePluginMappings(cfg *config.HarnessConfig) ([]adapter.Fi
 	return files, nil
 }
 
-func (a *Adapter) renderRouterSkill(cfg *config.HarnessConfig) (string, error) {
-	tmplContent, err := templates.FS.ReadFile("claude/commands/auto-router.md.tmpl")
-	if err != nil {
-		return "", fmt.Errorf("codex router skill 템플릿 읽기 실패: %w", err)
-	}
-
-	rendered, err := a.engine.RenderString(string(tmplContent), cfg)
-	if err != nil {
-		return "", fmt.Errorf("codex router skill 템플릿 렌더링 실패: %w", err)
-	}
-
-	frontmatter, body := splitSkillFrontmatter(rendered)
-	body = pkgcontent.ReplacePlatformReferences(body, "codex")
-	body = strings.Replace(body, "# /auto — Autopus-ADK", "# auto — Autopus-ADK", 1)
-	body = strings.Replace(body, "ARGUMENTS: $ARGUMENTS", "Interpret the user's text after the explicit `auto` mention as the command arguments. Support both `@auto ...` via the local Codex plugin and `$auto ...` via the repository skill.", 1)
-	body = normalizeCodexInvocationBody(body)
-	body = normalizeCodexHelperPaths(body)
-	body = normalizeCodexToolingBody(body)
-	body = rewriteCodexTriageElicitation(body)
-	body = rewriteCodexTeamModeSection(body)
-
-	aliasNote := strings.TrimSpace(fmt.Sprintf(`
-## Codex Invocation
-
-Use this skill through either of these surfaces:
-
-- %s — when the local Autopus plugin is installed from %s
-- %s — when using repository-scoped skills only
-
-Throughout this document, prefer %s. If the plugin is not installed, the same examples can be invoked with %s.
-
-Where the original Autopus workflow mentions Claude-specific capabilities:
-
-- %s means ask the user directly in one concise message unless an equivalent native Codex UI exists.
-- %s maps to Codex subagent spawning via %s.
-- %s maps to %s when coordinating an already-running agent.
-- %s is not available in Codex; keep %s as a reserved compatibility flag for future native multi-agent support and use %s subagents as the default workflow today.
-- Detailed helper references live under %s and %s in this repository.
-`,
-		"`@auto <subcommand> ...`",
-		"`/plugins`",
-		"`$auto <subcommand> ...`",
-		"`@auto ...`",
-		"`$auto ...`",
-		"`AskUserQuestion`",
-		"`Agent(...)`",
-		"`spawn_agent(...)`",
-		"`SendMessage`",
-		"`send_input(...)`",
-		"`TeamCreate`",
-		"`--team`",
-		"`spawn_agent(...)`",
-		"`.codex/skills/`",
-		"`.codex/rules/autopus/`",
-	))
-
-	body = injectAfterFirstHeading(body, aliasNote)
-	if frontmatter == "" {
-		frontmatter = strings.TrimSpace(fmt.Sprintf(`---
-name: auto
-description: >
-  Autopus Codex router skill. Use when the user wants %s or %s workflows such as plan, go, fix, review, sync, canary, and idea.
----`, "`@auto ...`", "`$auto ...`"))
-	}
-
-	return frontmatter + "\n\n" + strings.TrimSpace(body) + "\n", nil
-}
-
-func (a *Adapter) renderTemplateAsSkill(cfg *config.HarnessConfig, spec standardSkillSpec) (string, error) {
-	tmplContent, err := templates.FS.ReadFile(spec.TemplatePath)
-	if err != nil {
-		return "", fmt.Errorf("codex skill 템플릿 읽기 실패 %s: %w", spec.TemplatePath, err)
-	}
-
-	rendered, err := a.engine.RenderString(string(tmplContent), cfg)
-	if err != nil {
-		return "", fmt.Errorf("codex skill 템플릿 렌더링 실패 %s: %w", spec.Name, err)
-	}
-
-	_, body := splitSkillFrontmatter(rendered)
-	if strings.TrimSpace(body) == "" {
-		body = rendered
-	}
-
-	body = strings.TrimSpace(body)
-	body = pkgcontent.ReplacePlatformReferences(body, "codex")
-	body = normalizeCodexSkillBody(body, strings.TrimPrefix(spec.Name, "auto-"))
-	if !strings.Contains(body, "## Codex Invocation") {
-		invocationNote := strings.TrimSpace(fmt.Sprintf(`
-## Codex Invocation
-
-You can invoke this workflow through any of these compatible surfaces:
-
-- %s — preferred when the local Autopus plugin is installed
-- %s — direct repository skill invocation
-- %s — via the router skill
-
-Load and follow any helper documents referenced from this file under %s and %s.
-`,
-			fmt.Sprintf("`@auto %s ...`", strings.TrimPrefix(spec.Name, "auto-")),
-			fmt.Sprintf("`$%s ...`", spec.Name),
-			fmt.Sprintf("`$auto %s ...`", strings.TrimPrefix(spec.Name, "auto-")),
-			"`.codex/skills/`",
-			"`.codex/rules/autopus/`",
-		))
-		body = injectAfterFirstHeading(body, invocationNote)
-	}
-
-	frontmatter := fmt.Sprintf("---\nname: %s\ndescription: >\n  %s\n---", spec.Name, spec.Description)
-	return frontmatter + "\n\n" + body + "\n", nil
-}
-
-func normalizeCodexSkillBody(body, subcommand string) string {
-	body = normalizeCodexInvocationBody(body)
-	body = normalizeCodexHelperPaths(body)
-	body = normalizeCodexToolingBody(body)
-	if subcommand == "" {
-		return body
-	}
-
-	replacer := strings.NewReplacer(
-		fmt.Sprintf("@auto-%s", subcommand), fmt.Sprintf("@auto %s", subcommand),
-		fmt.Sprintf("$auto-%s", subcommand), fmt.Sprintf("$auto %s", subcommand),
-	)
-	return replacer.Replace(body)
-}
-
-func normalizeCodexInvocationBody(body string) string {
-	replacer := strings.NewReplacer(
-		"`/auto ", "`@auto ",
-		"/auto ", "@auto ",
-		"`/auto`", "`@auto`",
-	)
-	return replacer.Replace(body)
-}
-
-func normalizeCodexHelperPaths(body string) string {
-	replacer := strings.NewReplacer(
-		"@.codex/skills/autopus/", "@.codex/skills/",
-		".codex/skills/autopus/", ".codex/skills/",
-		"@.claude/skills/autopus/", "@.codex/skills/",
-		".claude/skills/autopus/", ".codex/skills/",
-		".codex/agents/autopus/", ".codex/agents/",
-		".claude/agents/autopus/", ".codex/agents/",
-		".claude/rules/autopus/", ".codex/rules/autopus/",
-	)
-	return replacer.Replace(body)
-}
-
-func normalizeCodexToolingBody(body string) string {
-	replacer := strings.NewReplacer(
-		"Load the `mcp__sequential-thinking__sequentialthinking` tool via ToolSearch, then perform step-by-step reasoning.", "Use sequential-thinking tooling if available; otherwise perform explicit step-by-step reasoning in the main Codex session.",
-		"Load the `WebSearch-thinking__sequentialthinking` tool via ToolSearch, then perform step-by-step reasoning.", "Use sequential-thinking tooling if available; otherwise perform explicit step-by-step reasoning in the main Codex session.",
-		"Use TeamCreate to create a team, then spawn specialized teammates using `Agent(subagent_type=..., team_name=..., name=...)`. Each teammate loads its agent definition from `.codex/agents/`, inheriting tools, skills, model, and domain expertise. Teammates communicate directly via SendMessage.", "Spawn specialized agents with `spawn_agent(...)` and coordinate them from the main session with `send_input(...)` and `wait_agent(...)`. Assign each worker a disjoint write scope and do not rely on Claude Code Team APIs.",
-		"For parallel tasks, include `auto pipeline worktree` in Agent() calls to enable worktree isolation.", "For parallel tasks, spawn separate workers with disjoint write scopes and follow `.codex/skills/worktree-isolation.md` for branch-isolation guidance.",
-		"For parallel tasks, include `auto pipeline worktree` in spawn_agent() calls to enable worktree isolation.", "For parallel tasks, spawn separate workers with disjoint write scopes and follow `.codex/skills/worktree-isolation.md` for branch-isolation guidance.",
-		"Claude Code automatically creates a worktree when `auto pipeline worktree` is passed to Agent(). No manual `git worktree add` is needed.", "Codex should not assume implicit worktree provisioning from agent flags. Use the documented worktree-isolation procedure when branch separation is required.",
-		"Each Phase below MUST use an Agent() call", "Each Phase below MUST use a `spawn_agent(...)` call",
-		"using the Agent tool", "using the `spawn_agent` tool",
-		"maps to Codex subagent spawning.", "maps to Codex subagent spawning.",
-		"Phase 0.5: Permission    → detect      (auto permission detect)", "Phase 0.5: Permission    → main session (decide autonomy vs confirmation)",
-	)
-	body = replacer.Replace(body)
-	body = strings.ReplaceAll(body, "Agent(", "spawn_agent(")
-	body = strings.ReplaceAll(body, "subagent_type =", "agent_type =")
-	body = strings.ReplaceAll(body, "prompt = ", "message = ")
-	body = strings.ReplaceAll(body, " (parallel, mode: plan)", " (parallel)")
-	body = strings.ReplaceAll(body, " (mode: plan)", "")
-	body = strings.ReplaceAll(body, " (mode: bypassPermissions)", "")
-	body = strings.ReplaceAll(body, " (mode: bypassPermissions, parallel with worktree isolation)", " (parallel with worktree isolation)")
-	body = strings.ReplaceAll(body, "  mode = PERMISSION_MODE == \"bypass\" ? \"bypassPermissions\" : \"plan\",\n", "")
-	body = strings.ReplaceAll(body, "  mode = \"bypassPermissions\",\n", "")
-	body = strings.ReplaceAll(body, "  mode = \"plan\",\n", "")
-	body = strings.ReplaceAll(body, "    permissionMode = \"bypassPermissions\",\n", "")
-	body = strings.ReplaceAll(body, "    permissionMode = \"plan\",\n", "")
-	body = strings.ReplaceAll(body, "  permissionMode = \"bypassPermissions\",\n", "")
-	body = strings.ReplaceAll(body, "  permissionMode = \"plan\",\n", "")
-	body = strings.ReplaceAll(body, "PERMISSION_MODE=$(auto permission detect)\n", "")
-	body = strings.ReplaceAll(body, "If the command fails or is unavailable, default to `PERMISSION_MODE=\"safe\"`.\n", "")
-	return body
-}
-
-func rewriteCodexTriageElicitation(body string) string {
-	replacement := "2. Ask the user directly in one concise message when confirmation is required. Present the recommended flow first and keep the choice in plain text; do not rely on unavailable UI-only tools.\n\n" +
-		"Example prompt:\n\n" +
-		"```\n" +
-		"추천 플로우는 `@auto idea ... --multi` 입니다. 이 플로우로 진행할까요, 아니면 `@auto fix` / `@auto plan` 중 다른 플로우로 바꿀까요?\n" +
-		"```"
-
-	return replaceCodexSectionInclusive(body,
-		"2. Use the `AskUserQuestion` tool to present the selection (do NOT use text-based numbered options):",
-		"Adjust the recommended option (add \"(Recommended)\" suffix) based on the difficulty classification. Place the recommended option FIRST in the options list.",
-		replacement,
-	)
-}
-
-func rewriteCodexTeamModeSection(body string) string {
-	replacement := `#### Route B: Reserved Team Flag (` + "`--team`" + `)
-
-IMPORTANT: Codex does not currently provide a documented native Team API equivalent to Claude Code ` + "`TeamCreate`" + `.
-
-When ` + "`--team`" + ` is set in Codex today:
-- Treat it as a reserved compatibility flag for future native multi-agent support.
-- Continue with Route A's default ` + "`spawn_agent(...)`" + ` subagent pipeline.
-- Do not reinterpret the flag as extra fan-out or a special harness-only worker topology.
-- Keep ownership isolation and validation rules identical to Route A.
-
-Revisit this route only when Codex exposes a documented native multi-agent surface that is distinct from ordinary subagent spawning.
-
-`
-
-	return replaceCodexSection(body,
-		"#### Route B: Agent Teams (`--team`)",
-		"#### Route C: Single Session (`--solo`)",
-		replacement,
-	)
-}
-
-func replaceCodexSection(body, start, end, replacement string) string {
-	startIdx := strings.Index(body, start)
-	if startIdx == -1 {
-		return body
-	}
-	endIdx := strings.Index(body[startIdx:], end)
-	if endIdx == -1 {
-		return body
-	}
-	endIdx += startIdx
-	return body[:startIdx] + replacement + body[endIdx:]
-}
-
-func replaceCodexSectionInclusive(body, start, end, replacement string) string {
-	startIdx := strings.Index(body, start)
-	if startIdx == -1 {
-		return body
-	}
-	endIdx := strings.Index(body[startIdx:], end)
-	if endIdx == -1 {
-		return body
-	}
-	endIdx += startIdx + len(end)
-	return body[:startIdx] + replacement + body[endIdx:]
-}
-
-func (a *Adapter) renderPluginManifestJSON() (string, error) {
-	doc := pluginManifest{
-		Name:        "auto",
-		Version:     "1.0.0",
-		Description: "Autopus workflow router for Codex: plan, go, fix, review, sync, canary, and idea.",
-		Author: pluginAuthor{
-			Name:  "Autopus",
-			Email: "noreply@autopus.co",
-			URL:   "https://autopus.co",
-		},
-		Homepage:   "https://autopus.co",
-		Repository: "https://github.com/insajin/autopus-adk",
-		License:    "Apache-2.0",
-		Keywords:   []string{"autopus", "workflow", "planning", "codex", "multi-provider"},
-		Skills:     "./skills",
-		Interface: pluginInterface{
-			DisplayName:       "Auto",
-			ShortDescription:  "Autopus workflow router for Codex",
-			LongDescription:   "Run Autopus plan/go/fix/review/sync/canary/idea workflows from Codex with a local plugin plus repository-managed helper docs.",
-			DeveloperName:     "Autopus",
-			Category:          "Developer Tools",
-			Capabilities:      []string{"Interactive", "Write", "Planning"},
-			WebsiteURL:        "https://autopus.co",
-			PrivacyPolicyURL:  "https://autopus.co/privacy",
-			TermsOfServiceURL: "https://autopus.co/terms",
-			DefaultPrompt: []string{
-				"@auto plan \"새 기능 요구사항을 SPEC으로 정리해줘\"",
-				"@auto go SPEC-EXAMPLE-001",
-				"@auto idea \"새 워크플로우를 멀티 프로바이더로 토론해줘\" --multi",
-			},
-			BrandColor: "#0F766E",
-		},
-	}
-
-	data, err := json.MarshalIndent(doc, "", "  ")
-	if err != nil {
-		return "", fmt.Errorf("plugin.json 직렬화 실패: %w", err)
-	}
-
-	return string(data) + "\n", nil
-}
-
-func (a *Adapter) renderMarketplaceJSON() (string, error) {
-	doc := marketplaceDoc{
-		Name: "autopus-local",
-		Interface: marketplaceUI{
-			DisplayName: "Autopus Local",
-		},
-		Plugins: []marketplaceEntry{
-			{
-				Name: "auto",
-				Source: marketplaceSource{
-					Source: "local",
-					Path:   "./.autopus/plugins/auto",
-				},
-				Policy: marketplacePolicy{
-					Installation:   "AVAILABLE",
-					Authentication: "ON_INSTALL",
-				},
-				Category: "Developer Tools",
-			},
-		},
-	}
-
-	existingPath := filepath.Join(a.root, ".agents", "plugins", "marketplace.json")
-	if data, err := os.ReadFile(existingPath); err == nil {
-		var existing marketplaceDoc
-		if jsonErr := json.Unmarshal(data, &existing); jsonErr == nil {
-			if existing.Name != "" {
-				doc.Name = existing.Name
-			}
-			if existing.Interface.DisplayName != "" {
-				doc.Interface.DisplayName = existing.Interface.DisplayName
-			}
-			updated := false
-			for i := range existing.Plugins {
-				if existing.Plugins[i].Name == "auto" {
-					existing.Plugins[i] = doc.Plugins[0]
-					updated = true
-					break
-				}
-			}
-			if !updated {
-				existing.Plugins = append(existing.Plugins, doc.Plugins[0])
-			}
-			doc.Plugins = existing.Plugins
-		}
-	}
-
-	data, err := json.MarshalIndent(doc, "", "  ")
-	if err != nil {
-		return "", fmt.Errorf("marketplace.json 직렬화 실패: %w", err)
-	}
-
-	return string(data) + "\n", nil
-}
-
 func newSkillMapping(targetPath, content string) adapter.FileMapping {
 	return adapter.FileMapping{
 		TargetPath:      targetPath,
@@ -566,36 +160,4 @@ func newSkillMapping(targetPath, content string) adapter.FileMapping {
 		Checksum:        checksum(content),
 		Content:         []byte(content),
 	}
-}
-
-func splitSkillFrontmatter(content string) (string, string) {
-	if !strings.HasPrefix(content, "---\n") {
-		return "", strings.TrimSpace(content)
-	}
-
-	rest := strings.TrimPrefix(content, "---\n")
-	idx := strings.Index(rest, "\n---\n")
-	if idx < 0 {
-		return "", strings.TrimSpace(content)
-	}
-
-	frontmatter := strings.TrimSpace(content[:len("---\n")+idx+len("\n---")])
-	body := strings.TrimSpace(rest[idx+len("\n---\n"):])
-	return frontmatter, body
-}
-
-func injectAfterFirstHeading(body, block string) string {
-	lines := strings.Split(body, "\n")
-	for i, line := range lines {
-		if strings.HasPrefix(line, "# ") {
-			out := make([]string, 0, len(lines)+4)
-			out = append(out, lines[:i+1]...)
-			out = append(out, "")
-			out = append(out, block)
-			out = append(out, "")
-			out = append(out, lines[i+1:]...)
-			return strings.Join(out, "\n")
-		}
-	}
-	return block + "\n\n" + body
 }
