+++ pkg/adapter/opencode/opencode_skills.go
@@ -3,6 +3,7 @@ package opencode
 import (
 	"fmt"
 	"path/filepath"
+	"strings"
 
 	contentfs "github.com/insajin/autopus-adk/content"
 	"github.com/insajin/autopus-adk/pkg/adapter"
@@ -26,22 +27,25 @@ func (a *Adapter) prepareSkillMappings(cfg *config.HarnessConfig) ([]adapter.Fil
 func (a *Adapter) prepareWorkflowSkillMappings(cfg *config.HarnessConfig) ([]adapter.FileMapping, error) {
 	files := make([]adapter.FileMapping, 0, len(workflowSpecs))
 	for _, spec := range workflowSpecs {
-		rendered, err := a.renderWorkflowPrompt(spec.PromptPath, cfg)
+		rendered, err := a.renderWorkflowSkill(spec, cfg)
 		if err != nil {
 			return nil, err
 		}
-		_, body := splitFrontmatter(rendered)
-		body = normalizeOpenCodeMarkdown(body)
-		body = skillInvocationNote(spec.Name) + "\n" + body
-		content := buildMarkdown(fmt.Sprintf("name: %s\ndescription: %q\ncompatibility: opencode", spec.Name, spec.Description), body)
 		files = append(files, adapter.FileMapping{
 			TargetPath:      filepath.Join(".agents", "skills", spec.Name, "SKILL.md"),
 			OverwritePolicy: adapter.OverwriteAlways,
-			Checksum:        adapter.Checksum(content),
-			Content:         []byte(content),
+			Checksum:        adapter.Checksum(rendered),
+			Content:         []byte(rendered),
 		})
 	}
 	return files, nil
+}
+
+func (a *Adapter) renderWorkflowSkill(spec workflowSpec, cfg *config.HarnessConfig) (string, error) {
+	if spec.Name == "auto" {
+		return a.renderRouterSkill(cfg)
+	}
+	return a.renderTemplateAsSkill(cfg, spec)
 }
 
 func (a *Adapter) prepareExtendedSkillMappings() ([]adapter.FileMapping, error) {
@@ -80,4 +84,46 @@ func (a *Adapter) renderWorkflowPrompt(templatePath string, cfg *config.HarnessC
 		return "", fmt.Errorf("workflow 템플릿 렌더링 실패 %s: %w", templatePath, err)
 	}
 	return rendered, nil
+}
+
+func (a *Adapter) renderRouterSkill(cfg *config.HarnessConfig) (string, error) {
+	rendered, err := a.renderWorkflowPrompt("codex/prompts/auto.md.tmpl", cfg)
+	if err != nil {
+		return "", err
+	}
+
+	_, body := splitFrontmatter(rendered)
+	if strings.TrimSpace(body) == "" {
+		body = rendered
+	}
+
+	body = strings.TrimSpace(body)
+	body = normalizeOpenCodeMarkdown(strings.TrimSpace(body))
+	body = skillInvocationNote("auto") + "\n" + body
+	body = body + "\n\n이 스킬은 얇은 라우터입니다. 서브커맨드를 해석한 뒤에는 반드시 대응하는 상세 스킬(`auto-plan`, `auto-go`, `auto-fix`, `auto-review`, `auto-sync`, `auto-canary`, `auto-idea`)을 추가로 로드해 실제 단계를 따르세요."
+
+	frontmatter := fmt.Sprintf("name: %s\ndescription: %q\ncompatibility: opencode", "auto", "Autopus 명령 라우터 — plan/go/fix/review/sync/canary/idea 서브커맨드를 해석합니다")
+	return buildMarkdown(frontmatter, body), nil
+}
+
+func (a *Adapter) renderTemplateAsSkill(cfg *config.HarnessConfig, spec workflowSpec) (string, error) {
+	rendered, err := a.renderWorkflowPrompt(spec.SkillPath, cfg)
+	if err != nil {
+		return "", err
+	}
+
+	_, body := splitFrontmatter(rendered)
+	if strings.TrimSpace(body) == "" {
+		body = rendered
+	}
+
+	body = strings.TrimSpace(body)
+	body = pkgcontent.ReplacePlatformReferences(body, "opencode")
+	body = normalizeOpenCodeSkillBody(body, strings.TrimPrefix(spec.Name, "auto-"))
+	if !strings.Contains(body, "## OpenCode Invocation") {
+		body = injectAfterFirstHeading(body, strings.TrimSpace(skillInvocationNote(spec.Name)))
+	}
+
+	frontmatter := fmt.Sprintf("name: %s\ndescription: %q\ncompatibility: opencode", spec.Name, spec.Description)
+	return buildMarkdown(frontmatter, body), nil
 }
