+++ templates/gemini/skills/auto-canary/SKILL.md.tmpl
@@ -1,0 +1,51 @@
+---
+name: auto-canary
+version: "1.0.0"
+description: 배포 검증 — build, E2E, 브라우저 건강 검진을 실행합니다
+platform: gemini-cli
+project: {{.ProjectName}}
+mode: {{.Mode}}
+triggers:
+  - "canary"
+  - "health check"
+  - "배포 검증"
+---
+
+# auto-canary — 배포 검증
+
+**프로젝트**: {{.ProjectName}} | **모드**: {{.Mode}}
+
+## 설명
+
+배포 후 건강 검진을 실행합니다. 빌드, E2E, 브라우저 검사를 순서대로 수행하고 PASS/WARN/FAIL 판정을 냅니다.
+
+## 사용법
+
+```bash
+@auto-canary
+@auto-canary --url https://example.com
+@auto-canary --watch 5m
+@auto-canary --compare abc123
+```
+
+## 플래그
+
+| Flag | Description |
+|------|-------------|
+| `--url <url>` | 브라우저 검진 대상 URL. `canary.md`의 browser 타깃보다 우선합니다. |
+| `--watch <interval>` | 지정 주기로 반복 실행합니다. 기본 5m, 최대 30m. |
+| `--compare <commit>` | 저장된 이전 실행 결과와 비교합니다. |
+
+## 실행 순서
+
+1. `.autopus/project/canary.md` 또는 `.autopus/project/tech.md` 기준으로 빌드 검증
+2. `.autopus/project/scenarios.md` 기준으로 E2E 실행
+3. 브라우저 건강 검진 수행
+4. PASS/WARN/FAIL 판정
+5. `.autopus/canary/latest.json`에 결과 저장
+
+## 판정 기준
+
+- **PASS**: 빌드 OK + 전체 E2E 통과 + 치명적 브라우저 오류 없음
+- **WARN**: 빌드 OK + 일부 E2E 실패 또는 비치명적 브라우저 경고
+- **FAIL**: 빌드 실패 또는 치명적 E2E/브라우저 오류
